Corrected Kendra FAQ Handling for the Test feature
